autoy/text2video/requirements.txt
gTTS>=2.3.2
moviepy>=1.0.3
numpy>=1.25.2
opencv-python>=4.8.0.74
pandas>=2.0.3
Pillow>=10.0.0